requirements.txt
flask-restplus==0.9.2
Flask-SQLAlchemy==2.1
Werkzeug==0.16.1
Authlib==0.14.3
requests==2.24.0